IN-111 fix: form styles and imports
