+++ src/features/profile/settings/generalInfo/api/profile.api.ts
@@ -1,26 +1,9 @@
 import { Endpoints, instance } from '@/features/auth'
-import { PhotoResult } from '@/features/profile/settings/avatarDialog/model'
-import { FormData as FormDataType } from '@/features/profile/settings/generalInfo/model/useFillGeneralForm'
+import {
+  UpdatedProfile,
+  UserProfile,
+} from '@/features/profile/settings/generalInfo/api'
 import { AxiosInstance, AxiosResponse } from 'axios'
-
-export type UpdatedProfile = {
-  photoData?: PhotoResult | undefined
-  region: string
-} & Omit<Required<FormDataType>, 'photoData'>
-
-export type UserProfile = {
-  avatars: Avatar[]
-  createdAt: string
-  id: number
-} & UpdatedProfile
-
-export type Avatar = {
-  createdAt: string
-  fileSize: number
-  height: number
-  url: string
-  width: number
-}
 
 class ProfileApi {
   constructor(private instance: AxiosInstance) {}
