+++ src/features/profile/settings/generalInfo/index.ts
@@ -1,0 +1,3 @@
+export * from './api'
+export * from './model'
+export * from './ui'
