+++ src/features/profile/settings/generalInfo/model/index.ts
@@ -1,0 +1,4 @@
+export * from './generalInfoSchema'
+export * from './useFetchLocations'
+export * from './useFillGeneralForm'
+export * from './useHandleCalendar'
