+++ src/features/profile/settings/generalInfo/ui/index.tsx
@@ -35,11 +35,11 @@ export const GeneralInfoForm = observer(
     return (
       <div
         className={
-          'flex gap-10 w-full mt-6 relative after:absolute after:contain-content after:h-[1px] after:top-[90%] after:left-0 after:w-full after:bg-dark-300'
+          'flex w-full mt-6 relative after:absolute after:contain-content after:h-[1px] after:top-[90%] after:left-0 after:w-full after:bg-dark-300'
         }
       >
         <form
-          className={'w-full flex flex-col gap-6'}
+          className={'flex gap-10 w-full'}
           onSubmit={handleSubmit(onSubmit)}
           ref={ref}
         >
@@ -47,104 +47,106 @@ export const GeneralInfoForm = observer(
             control={control}
             name={'photoData'}
           />
-          <FormTextField
-            control={control}
-            label={'Username'}
-            name={'userName'}
-            required
-          />
-          <FormTextField
-            control={control}
-            label={'First Name'}
-            name={'firstName'}
-            required
-          />
-          <FormTextField
-            control={control}
-            label={'Last Name'}
-            name={'lastName'}
-            required
-          />
-          <div className={'flex flex-col'}>
+          <div className={'w-full flex flex-col gap-6'}>
             <FormTextField
               control={control}
-              label={'Date of Birth'}
-              name={'dateOfBirth'}
-              readOnly
-            >
-              {!isCalendarOpen ? (
-                <CalendarOutline
-                  className={
-                    'absolute -translate-y-1/2 top-1/2 stroke-width-1 fill-light-100 right-3 cursor-pointer w-[18px] h-5'
-                  }
-                  onClick={toggleCalendar}
-                />
-              ) : (
-                <CalendarFilled
-                  className={
-                    'absolute -translate-y-1/2 top-1/2 stroke-width-1 fill-light-100 right-3 cursor-pointer w-[18px] h-5'
-                  }
-                  onClick={toggleCalendar}
-                />
-              )}
-              {isCalendarOpen && (
-                <div ref={calendarRef}>
-                  <Calendar
-                    className={'absolute'}
-                    onDayClick={selectDateHandler}
+              label={'Username'}
+              name={'userName'}
+              required
+            />
+            <FormTextField
+              control={control}
+              label={'First Name'}
+              name={'firstName'}
+              required
+            />
+            <FormTextField
+              control={control}
+              label={'Last Name'}
+              name={'lastName'}
+              required
+            />
+            <div className={'flex flex-col'}>
+              <FormTextField
+                control={control}
+                label={'Date of Birth'}
+                name={'dateOfBirth'}
+                readOnly
+              >
+                {!isCalendarOpen ? (
+                  <CalendarOutline
+                    className={
+                      'absolute -translate-y-1/2 top-1/2 stroke-width-1 fill-light-100 right-3 cursor-pointer w-[18px] h-5'
+                    }
+                    onClick={toggleCalendar}
                   />
-                </div>
-              )}
-            </FormTextField>
-            <div className={'flex gap-6'}>
-              <FormSelect
-                className={'w-full'}
-                control={control}
-                label={'Select your country'}
-                name={'country'}
-                placeholder={'Country'}
-              >
-                {countriesData?.map((country) => (
-                  <SelectItem
-                    key={country.country}
-                    value={country.country}
-                  >
-                    {country.country}
-                  </SelectItem>
-                ))}
-              </FormSelect>
-              <FormSelect
-                className={'w-full'}
-                control={control}
-                disabled={!countryValue}
-                label={'Select your city'}
-                name={'city'}
-                placeholder={'City'}
-              >
-                {cities?.map((city) => (
-                  <SelectItem
-                    key={city}
-                    value={city}
-                  >
-                    {city}
-                  </SelectItem>
-                ))}
-              </FormSelect>
+                ) : (
+                  <CalendarFilled
+                    className={
+                      'absolute -translate-y-1/2 top-1/2 stroke-width-1 fill-light-100 right-3 cursor-pointer w-[18px] h-5'
+                    }
+                    onClick={toggleCalendar}
+                  />
+                )}
+                {isCalendarOpen && (
+                  <div ref={calendarRef}>
+                    <Calendar
+                      className={'absolute'}
+                      onDayClick={selectDateHandler}
+                    />
+                  </div>
+                )}
+              </FormTextField>
+              <div className={'flex gap-6'}>
+                <FormSelect
+                  className={'w-full'}
+                  control={control}
+                  label={'Select your country'}
+                  name={'country'}
+                  placeholder={'Country'}
+                >
+                  {countriesData?.map((country) => (
+                    <SelectItem
+                      key={country.country}
+                      value={country.country}
+                    >
+                      {country.country}
+                    </SelectItem>
+                  ))}
+                </FormSelect>
+                <FormSelect
+                  className={'w-full'}
+                  control={control}
+                  disabled={!countryValue}
+                  label={'Select your city'}
+                  name={'city'}
+                  placeholder={'City'}
+                >
+                  {cities?.map((city) => (
+                    <SelectItem
+                      key={city}
+                      value={city}
+                    >
+                      {city}
+                    </SelectItem>
+                  ))}
+                </FormSelect>
+              </div>
             </div>
-          </div>
-          <FormTextArea
-            control={control}
-            label={'About me'}
-            name={'aboutMe'}
-          />
-          <div className={'flex justify-end mt-12'}>
-            <Button
-              disabled={!isValid || isSubmitting}
-              type={'submit'}
-              variant={'primary'}
-            >
-              Save Changes
-            </Button>
+            <FormTextArea
+              control={control}
+              label={'About me'}
+              name={'aboutMe'}
+            />
+            <div className={'flex justify-end mt-12'}>
+              <Button
+                disabled={!isValid || isSubmitting}
+                type={'submit'}
+                variant={'primary'}
+              >
+                Save Changes
+              </Button>
+            </div>
           </div>
         </form>
         <ToastContainer />
